- Changed DDS Module amplitude in firmware + software value to 16-bit, incremented firmware version to 2 - Changed BpodSettings.mat example settings file to a different name. If the user accidentally adds all subfolders to path, this change prevents the example settings from being used.

+++ Functions/Modules/DDS/DDSModule.m
@@ -105,7 +105,7 @@ classdef DDSModule < handle
                     error('Error: amplitude must be in range [0,1]')
                 end
                 ampValue = floor(amp*10000);
-                obj.Port.write('A', 'uint8', ampValue, 'uint32');
+                obj.Port.write('A', 'uint8', ampValue, 'uint16');
                 Confirmed = obj.Port.read(1, 'uint8');
                 if Confirmed ~= 1
                     error('Error setting amplitude. Confirm code not returned.');
